requirements.txt
Django==3.1.5
djangorestframework==3.12.2
django-rest-knox==0.1.11
tika==1.24
